NeuroCrowd/requirements.txt
mesa==2.1.2
streamlit==1.32.0
plotly==5.18.0
numpy==1.26.4
python-dotenv==1.0.1
google-generativeai==0.3.2